Fix build warnings in src/xrGame

+++ src/xrGame/ef_pattern.h
@@ -44,7 +44,7 @@ private:
     {
         SPattern& tPattern = m_tpPatterns[iPatternIndex];
         u32 dwIndex = dwpTest[tPattern.dwaVariableIndexes[0]];
-        for (u32 i = 1; i < (int)tPattern.dwCardinality; ++i)
+        for (u32 i = 1; i < tPattern.dwCardinality; ++i)
             dwIndex = dwIndex * m_dwaAtomicFeatureRange[tPattern.dwaVariableIndexes[i]] +
                 dwpTest[tPattern.dwaVariableIndexes[i]];
         return (dwIndex + m_dwaPatternIndexes[iPatternIndex]);
